添加socketpair测试用例
